requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
ffmpeg-python==0.2.0   # FFmpeg bindings
redis==5.0.1
celery==5.3.4          # for background jobs
python-dotenv==1.0.0   # env handling
sqlalchemy>=2.0.0
aiofiles>=23.0.0

# Development and testing extras
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
pre-commit>=3.5.0
